on: 修改model

+++ server/app/Controllers/Account.php
@@ -4,19 +4,19 @@
  * @Author: freeair
  * @Date: 2021-06-25 11:16:41
  * @LastEditors: freeair
- * @LastEditTime: 2022-05-07 23:21:39
+ * @LastEditTime: 2022-05-10 21:14:31
  */
 
 namespace App\Controllers;
 
 use App\Models\Account\AccountModel;
-use App\Models\Account\AvatarModel;
-use App\Models\Account\DeptModel;
-use App\Models\Account\JobModel;
-use App\Models\Account\MenuModel;
-use App\Models\Account\PoliticModel;
-use App\Models\Account\TitleModel;
-use App\Models\Account\UserModel;
+use App\Models\Common\AvatarModel;
+use App\Models\Common\DeptModel;
+use App\Models\Common\JobModel;
+use App\Models\Common\MenuModel;
+use App\Models\Common\PoliticModel;
+use App\Models\Common\TitleModel;
+use App\Models\Common\UserModel;
 use App\Models\SmsCodeModel;
 use CodeIgniter\API\ResponseTrait;
 
@@ -87,7 +87,14 @@ class Account extends BaseController
 
         // 取session的id
         $client = $this->request->getJSON(true);
+        $cid    = $client['id'];
         $sid    = $this->session->get('id');
+
+        if ($cid !== $sid) {
+            $res['info']  = 'invalid id';
+            $res['error'] = '请求数据无效';
+            return $this->fail($res);
+        }
 
         // 取出检验后的数据
         $newSetting = [
@@ -103,7 +110,7 @@ class Account extends BaseController
 
         // 修改数据库
         $model = new AccountModel();
-        $uid   = $model->updateUserInfo($newSetting);
+        $uid   = $model->updateUserSetting($newSetting);
         if (!$uid) {
             $phone = $this->session->get('phone');
             log_message('error', '{file}:{line} --> update user info db failed ' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
@@ -111,21 +118,7 @@ class Account extends BaseController
             return $this->failServerError('服务器处理发生错误，稍候再试');
         }
 
-        // 读数据库
-        $query = $this->getUserByQuery(['id' => $sid]);
-
-        $user = $query['result'][0];
-        // 取用户头像文件
-        if (isset($user['avatar']) && is_numeric($user['avatar'])) {
-            $avatarModel = new AvatarModel();
-            $avatar      = $avatarModel->getAvatarPathAndNameById($user['avatar']);
-            if (isset($avatar['path']) && isset($avatar['name'])) {
-                $user['avatarFile'] = $avatar['path'] . $avatar['name'];
-            }
-        }
-
-        // 修改session
-        $user['dept_ids'] = explode("+", trim($user['dept_ids'], '+'));
+        $user = $this->_preSetSession($sid);
         $this->session->set($user);
 
         $res['msg']  = '保存成功';
@@ -136,139 +129,101 @@ class Account extends BaseController
 
     public function updatePassword()
     {
-        // 检查请求数据
         if (!$this->validate('AccountUpdatePassword')) {
-            $res['error'] = $this->validator->getErrors();
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '请求数据无效！';
-            return $this->respond($res);
+            $res['info']  = $this->validator->getErrors();
+            $res['error'] = '请求数据无效';
+            return $this->fail($res);
         }
 
-        // 取session的id
         $client = $this->request->getJSON(true);
         $id     = $this->session->get('id');
 
-        // 取出检验后的数据
         $password    = $client['password'];
         $newPassword = $client['newPassword'];
 
         // 验证密码
         $userModel    = new UserModel();
-        $hashPassword = $userModel->getUserPassword(['id' => $id]);
+        $hashPassword = $userModel->getUserPwdByUid($id);
 
         $result = my_verify_password($password, $hashPassword);
         if ($result === false) {
             log_message('error', '{file}:{line} --> password error when update user password ');
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '旧登录密码错误！';
-            return $this->respond($res);
+            return $this->fail('旧登录密码错误');
         }
 
         // 修改数据库
         $model = new AccountModel();
         if (!$model->updatePasswordById($id, $newPassword)) {
-            // $phone = $this->session->get('phone');
             log_message('error', '{file}:{line} --> update user password db failed ' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '修改失败，稍后再试！';
-            return $this->respond($res);
+            return $this->failServerError('服务器处理发生错误，稍候再试');
         }
 
-        $res['code'] = EXIT_SUCCESS;
-        $res['msg']  = '修改成功！';
-
+        $res['msg'] = '修改成功';
         return $this->respond($res);
     }
 
     public function updatePhone()
     {
-        // 检查请求数据
         if (!$this->validate('AccountUpdatePhone')) {
-            $res['error'] = $this->validator->getErrors();
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '请求数据无效！';
-            return $this->respond($res);
+            $res['info']  = $this->validator->getErrors();
+            $res['error'] = '请求数据无效';
+            return $this->fail($res);
         }
 
-        // 取session的id
         $client = $this->request->getJSON(true);
         $id     = $this->session->get('id');
 
-        // 取出检验后的数据
         $password = $client['password'];
         $phone    = $client['phone'];
 
         // 检查手机号是否存在
         $userModel = new UserModel();
-        if ($userModel->getUserPassword(['phone' => $phone]) !== '') {
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '新手机号已被其他用户绑定！';
-            return $this->respond($res);
+        if ($userModel->getUserPwdByPhone($phone) !== '') {
+            return $this->fail('新手机号已被其他用户绑定');
         }
 
         // 验证密码
-        $hashPassword = $userModel->getUserPassword(['id' => $id]);
-
-        $result = my_verify_password($password, $hashPassword);
+        $hashPassword = $userModel->getUserPwdByUid($id);
+        $result       = my_verify_password($password, $hashPassword);
         if ($result === false) {
             log_message('error', '{file}:{line} --> password error when update user phone' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '密码错误！';
-            return $this->respond($res);
+            return $this->fail('密码错误！');
         }
 
         // 修改数据库
         $model = new AccountModel();
         if (!$model->updatePhoneById($id, $phone)) {
-            // $phone = $this->session->get('phone');
             log_message('error', '{file}:{line} --> update user phone db failed ' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
 
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '修改失败，稍后再试！';
-            return $this->respond($res);
+            return $this->failServerError('服务器处理发生错误，稍候再试');
         }
 
         // 修改session
         $this->session->set('phone', $phone);
 
-        $res['code'] = EXIT_SUCCESS;
-        $res['msg']  = '修改成功！';
+        $res['msg']  = '修改成功';
         $res['data'] = ['phone' => $phone];
-
         return $this->respond($res);
     }
 
     // 验证码
     public function sms()
     {
-        // 检查请求数据
         if (!$this->validate('AccountSMS')) {
-            $res['error'] = $this->validator->getErrors();
-            $res['code']  = EXIT_ERROR;
-            $res['msg']   = '请求数据无效！';
-            return $this->respond($res);
+            $res['info']  = $this->validator->getErrors();
+            $res['error'] = '请求数据无效';
+            return $this->fail($res);
         }
 
-        // 取session的id
         $client = $this->request->getJSON(true);
         $id     = $this->session->get('id');
 
-        // 取出请求数据
         $email = $client['email'];
 
         // 检查邮箱是否存在
         $userModel = new UserModel();
-        if ($userModel->getUserPassword(['email' => $email]) !== '') {
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '新邮箱已被其他用户绑定！';
-            return $this->respond($res);
+        if ($userModel->getUserPwdByEmail($email) !== '') {
+            return $this->fail('新邮箱已被其他用户绑定');
         }
 
         // 生成验证码
@@ -277,11 +232,7 @@ class Account extends BaseController
         $code         = $smsCodeModel->newSmsCodeByPhone($phone);
         if (empty($code)) {
             log_message('error', '{file}:{line} --> new sms code failed' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
-            $res['code']  = EXIT_ERROR;
-            $res['msg']   = '发送验证码失败，稍后尝试！';
-            $res['error'] = 'new sms code';
-            return $this->respond($res);
+            return $this->failServerError('服务器处理发生错误，稍候再试');
         }
 
         // 验证码发送邮件
@@ -299,46 +250,32 @@ class Account extends BaseController
         if (!$emailAPI->send(false)) {
             $err = $emailAPI->printDebugger('subject');
             log_message('error', '{file}:{line} --> send mail failed ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4) . '.  ' . $err);
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '发送验证码失败，稍后尝试！';
-            return $this->respond($res);
+            return $this->failServerError('服务器处理发生错误，稍候再试');
         }
 
-        // 屏蔽个人信息
-
-        $res['code'] = EXIT_SUCCESS;
         $res['data'] = ['email' => $email];
         return $this->respond($res);
     }
 
     public function updateEmail()
     {
-        // 检查请求数据
         if (!$this->validate('AccountUpdateEmail')) {
-            $res['error'] = $this->validator->getErrors();
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '请求数据无效！';
-            return $this->respond($res);
+            $res['info']  = $this->validator->getErrors();
+            $res['error'] = '请求数据无效';
+            return $this->fail($res);
         }
 
-        // 取session的id
         $client = $this->request->getJSON(true);
         $id     = $this->session->get('id');
 
-        // 取出检验后的数据
         $password = $client['password'];
         $email    = $client['email'];
         $code     = $client['code'];
 
         // 检查邮箱是否存在
         $userModel = new UserModel();
-        if ($userModel->getUserPassword(['email' => $email]) !== '') {
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '新邮箱已被其他用户绑定！';
-            return $this->respond($res);
+        if ($userModel->getUserPwdByEmail($email) !== '') {
+            return $this->fail('新邮箱已被其他用户绑定');
         }
 
         // 核对验证码
@@ -346,38 +283,29 @@ class Account extends BaseController
         $smsCodeModel = new SmsCodeModel();
         $isOK         = $smsCodeModel->validateSmsCodeByPhone($phone, $code);
         if (!$isOK) {
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '验证码无效！';
-            return $this->respond($res);
+            return $this->fail('验证码无效');
         }
 
         // 验证密码
-        $hashPassword = $userModel->getUserPassword(['id' => $id]);
+        $hashPassword = $userModel->getUserPwdByUid($id);
 
         $result = my_verify_password($password, $hashPassword);
         if ($result === false) {
             log_message('error', '{file}:{line} --> password error when update user email' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '密码错误！';
-            return $this->respond($res);
+            return $this->fail('密码错误');
         }
 
         // 修改数据库
         $model = new AccountModel();
         if (!$model->updateEmailById($id, $email)) {
             log_message('error', '{file}:{line} --> update user email db failed ' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '修改失败，稍后再试！';
-            return $this->respond($res);
+            return $this->failServerError('服务器处理发生错误，稍候再试');
         }
 
         // 修改session
         $this->session->set('email', $email);
 
-        $res['code'] = EXIT_SUCCESS;
-        $res['msg']  = '修改成功！';
+        $res['msg']  = '修改成功';
         $res['data'] = ['email' => $email];
 
         return $this->respond($res);
@@ -392,11 +320,7 @@ class Account extends BaseController
         // 检查文件
         if (!$file->isValid() || ($file->getSize() / 1024 / 1024 > 2) || $file->getMimeType() !== 'image/jpeg') {
             log_message('error', '{file}:{line} --> upload avatar file invalid ' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '头像图片文件超要求';
-            return $this->respond($res, 422);
-
+            return $this->fail('头像图片文件超要求');
         }
 
         $newName      = $file->getRandomName();
@@ -416,12 +340,9 @@ class Account extends BaseController
                 ->save($absoluteFile);
         } catch (\CodeIgniter\Images\ImageException $e) {
             log_message('error', '{file}:{line} --> upload avatar resize failed ' . $e->getMessage() . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
             // 删除文件
             unlink($absoluteFile);
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '头像图片文件调整失败';
-            return $this->respond($res, 422);
+            return $this->failServerError('头像图片文件调整失败');
         }
 
         // 查询头像文件路径和文件名，待后续删除旧头像文件
@@ -430,14 +351,11 @@ class Account extends BaseController
         $oldAvatar   = $avatarModel->getAvatarPathAndNameById($avatarId);
 
         // 更改数据库
-        if (!$avatarModel->updateAvatarById($avatarId, $relativePath, $newName)) {
+        if (!$avatarModel->updateAvatarById($relativePath, $newName, $avatarId)) {
             log_message('error', '{file}:{line} --> upload avatar db failed ' . '[' . substr(session_id(), 0, 15) . '] ' . substr($phone, 0, 3) . '****' . substr($phone, 7, 4));
-
             // 删除文件
             unlink($absoluteFile);
-            $res['code'] = EXIT_ERROR;
-            $res['msg']  = '修改DB失败';
-            return $this->respond($res, 422);
+            return $this->failServerError('服务器处理发生错误，稍候再试');
         }
 
         // 修改session
@@ -445,16 +363,16 @@ class Account extends BaseController
         $this->session->set('avatarFile', $avatarFile);
 
         // 删除旧文件
-        if (strpos($oldAvatar['path'], 'default') === false) {
-            $old = $absolutePath . $oldAvatar['name'];
-            if (file_exists($old)) {
-                unlink($absolutePath . $oldAvatar['name']);
+        if (!empty($oldAvatar)) {
+            if (strpos($oldAvatar['path'], 'default') === false) {
+                $old = $absolutePath . $oldAvatar['name'];
+                if (file_exists($old)) {
+                    unlink($absolutePath . $oldAvatar['name']);
+                }
             }
         }
 
-        $res['code']       = EXIT_SUCCESS;
         $res['avatarFile'] = $avatarFile;
-
         return $this->respond($res);
     }
 
@@ -515,7 +433,7 @@ class Account extends BaseController
         $fields = ['id', 'pid', 'name', 'path', 'component', 'redirect', 'hideChildrenInMenu', 'title', 'icon', 'keepAlive', 'meta_hidden', 'hiddenHeaderContent', 'hidden', 'permission', 'target'];
 
         $model = new MenuModel();
-        $db    = $model->getByPageIds($fields, $pageId);
+        $db    = $model->getMenuRecordsByIds($fields, $pageId);
         if (empty($db)) {
             return [];
         }
@@ -595,16 +513,16 @@ class Account extends BaseController
         ];
 
         $politicModel      = new politicModel();
-        $result['politic'] = $politicModel->getPolitic(['id', 'name', 'status']);
+        $result['politic'] = $politicModel->getPoliticAllRecords(['id', 'name', 'status']);
 
         $deptModel      = new DeptModel();
-        $result['dept'] = $deptModel->getDept(['id', 'name', 'pid', 'status']);
+        $result['dept'] = $deptModel->getDeptAllRecords(['id', 'name', 'pid', 'status']);
 
         $jobModel      = new JobModel();
-        $result['job'] = $jobModel->getJob(['id', 'name', 'status']);
+        $result['job'] = $jobModel->getJobAllRecords(['id', 'name', 'status']);
 
         $titleModel      = new titleModel();
-        $result['title'] = $titleModel->getTitle(['id', 'name', 'status']);
+        $result['title'] = $titleModel->getTitleAllRecords(['id', 'name', 'status']);
 
         $res['http_status'] = 200;
         $res['data']        = $result;
@@ -612,23 +530,78 @@ class Account extends BaseController
 
     }
 
-    protected function getUserByQuery($queryParam = [])
+    protected function _preSetSession(string $uid = null)
     {
-        $deptModel = new DeptModel();
-        $dept      = $deptModel->getDept(['id', 'name']);
+        if (empty($uid)) {
+            return [];
+        }
 
-        $jobModel = new JobModel();
-        $job      = $jobModel->getJob(['id', 'name']);
+        if (floor($uid) != $uid) {
+            return [];
+        }
 
-        $titleModel = new titleModel();
-        $title      = $titleModel->getTitle(['id', 'name']);
-
-        $politicModel = new politicModel();
-        $politic      = $politicModel->getPolitic(['id', 'name']);
-
+        $fields = ['id', 'workID', 'username', 'sex', 'IdCard', 'phone', 'email', 'status', 'avatar', 'dept_ids', 'job', 'title', 'politic', 'updated_at'];
         $model  = new UserModel();
-        $result = $model->getUserByQueryParam($dept, $job, $title, $politic, $queryParam);
+        $user   = $model->getUserRecordById($fields, $uid);
 
-        return $result;
+        if (empty($user)) {
+            return [];
+        }
+
+        $model = new DeptModel();
+        $dept  = $model->getDeptAllRecords(['id', 'name']);
+
+        $model = new JobModel();
+        $job   = $model->getJobAllRecords(['id', 'name']);
+
+        $model = new titleModel();
+        $title = $model->getTitleAllRecords(['id', 'name']);
+
+        $model   = new politicModel();
+        $politic = $model->getPoliticAllRecords(['id', 'name']);
+
+        // 部门名称
+        $department = '';
+        $deptIds    = explode("+", trim($user['dept_ids'], '+'));
+        $cnt        = count($deptIds);
+        for ($i = 0; $i < $cnt; $i++) {
+            foreach ($dept as $value) {
+                if ($value['id'] === $deptIds[$i]) {
+                    $department = $department . $value['name'] . ' / ';
+                }
+            }
+        }
+        $user['department'] = rtrim($department, " / ");
+        $user['dept_ids']   = $deptIds;
+
+        // 岗位名称
+        foreach ($job as $x) {
+            if ($user['job'] === $x['id']) {
+                $user['jobLabel'] = $x['name'];
+            }
+        }
+        // 职称名称
+        foreach ($title as $y) {
+            if ($user['title'] === $y['id']) {
+                $user['titleLabel'] = $y['name'];
+            }
+        }
+        // 政治面貌
+        foreach ($politic as $z) {
+            if ($user['politic'] === $z['id']) {
+                $user['politicLabel'] = $z['name'];
+            }
+        }
+
+        // 头像文件
+        if (isset($user['avatar']) && is_numeric($user['avatar'])) {
+            $avatarModel = new AvatarModel();
+            $avatar      = $avatarModel->getAvatarPathAndNameById($user['avatar']);
+            if (isset($avatar['path']) && isset($avatar['name'])) {
+                $user['avatarFile'] = $avatar['path'] . $avatar['name'];
+            }
+        }
+
+        return $user;
     }
 }
